requirements.txt
playwright==1.56.0
pandas==2.3.3
openpyxl==3.1.5
python-dotenv==1.1.1
pyinstaller==6.16.0